<feature> <mouse> : 增加鼠标检测功能

+++ kc_sdr/engine/engine.h
@@ -14,7 +14,7 @@ public:
     qint16          i16_simAmp;
     quint16         u16_simFreq;
 signals:
-    void fftGenerated(quint32 u32_addr, quint16 u16_size);
+    void fftGenerated(intptr_t u32_addr, quint16 u16_size);
 protected:
     void run();
 private:
